+++ src/utils/capitalizeFirstLetter.js
@@ -1,0 +1,7 @@
+export const capitalizeFirstLetter = (str, lowerRest = false) => {
+  const [first, ...rest] = str;
+  return (
+    first.toUpperCase() +
+    (lowerRest ? rest.join("").toLowerCase() : rest.join(""))
+  );
+};